test: Convert server tests to new test macros
