+++ Makefile.am
@@ -83,6 +83,7 @@ unit_test_LDADD = libtest.la
 integration_test_SOURCES = \
   test/integration/test_membership.c \
   test/integration/test_server.c \
+  test/integration/test_client.c \
   test/integration/main.c
 integration_test_CFLAGS = $(AM_CFLAGS)
 integration_test_LDFLAGS = $(AM_LDFLAGS) -no-install
